refactor: Move ValidationErrorHandler into AddFhir (#365)
* Moved the Validation Error Handler into AddFhir so it's part of
  Framework instead of a manual operation for users of the NuGet
  packages
