+++ src/Spark.Web/Startup.cs
@@ -19,7 +19,6 @@ using System.Linq;
 using Microsoft.Extensions.Logging;
 using Microsoft.AspNetCore.Mvc.Formatters;
 using Microsoft.Extensions.Hosting;
-using Hl7.Fhir.Rest;
 
 namespace Spark.Web
 {
@@ -112,28 +111,6 @@ namespace Spark.Web
             });
 
             services.AddSignalR();
-
-            // Make validation errors to be returned as application/json or application/xml
-            // instead of application/problem+json and application/problem+xml.
-            // (https://github.com/FirelyTeam/spark/issues/282)
-            services.Configure<ApiBehaviorOptions>(options =>
-            {
-                var defaultInvalidModelStateResponseFactory = options.InvalidModelStateResponseFactory;
-                options.InvalidModelStateResponseFactory = context =>
-                {
-                    var actionResult = defaultInvalidModelStateResponseFactory(context) as ObjectResult;
-                    if (actionResult != null)
-                    {
-                        actionResult.ContentTypes.Clear();
-                        foreach (var mediaType in ContentType.JSON_CONTENT_HEADERS
-                            .Concat(ContentType.XML_CONTENT_HEADERS))
-                        {
-                            actionResult.ContentTypes.Add(mediaType);
-                        }
-                    }
-                    return actionResult;
-                };
-            });
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
